pyproject.toml
[tool.poetry]
name = "django-rbaca"
version = "1.0.0"
description = "Role-based access control with attributes for Django"
authors = ["Steven Bruch", "Mickael Cormier"]
readme = "README.md"
packages = [{include = "rbaca"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.11"
djangorestframework = "^3.15.1"
drf-jwt = "^1.19.2"
sphinx = "^7.3.7"
sphinx-rtd-theme = "^1.3.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "24.3.0"
djlint = "^1.34.0"
isort = "^5.12.0"
safety = "^2.3.5"
flake8 = "^6.1.0"
coverage = "7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
